--- src/data_readers/unit_test/test_data/hdf5_data_schema.yaml
@@ -1,128 +1,0 @@
-std::string const data_schema = R"ptext(
-inputs: 
-  shape_model_initial_modes:(4,3): 
-    metadata: 
-      scale: 1.666672
-      bias: 0.5
-      ordering: 100
-  betti_prl15_trans_u: 
-    metadata: 
-      scale: 1.000002
-      bias: -1.603483e-07
-      ordering: 101
-  betti_prl15_trans_v: 
-    metadata: 
-      scale: 1.000001
-      bias: -1.406672e-06
-      ordering: 102
-  shape_model_initial_modes:(2,1): 
-    metadata: 
-      scale: 1.666675
-      bias: 0.4999992
-      ordering: 103
-  shape_model_initial_modes:(1,0): 
-    metadata: 
-      scale: 1.666669
-      bias: 0.5000008
-      ordering: 104
-outputs: 
-  scalars: 
-    BWx: 
-      metadata: 
-        scale: 7.610738
-        bias: -0.4075375
-        ordering: 201
-    BT: 
-      metadata: 
-        scale: 1.459875
-        bias: -3.427656
-        ordering: 202
-    tMAXt: 
-      metadata: 
-        scale: 1.490713
-        bias: -3.495498
-        ordering: 203
-    BWn: 
-      metadata: 
-        scale: 43.75123
-        bias: -1.593477
-        ordering: 204
-    MAXpressure: 
-      metadata: 
-        scale: 1.685576e-06
-        bias: -0.5330971
-        ordering: 205
-    tMAXpressure: 
-      metadata: 
-        scale: 1.430615
-        bias: -3.351173
-        ordering: 206
-    BAt: 
-      metadata: 
-        scale: 0.2636422
-        bias: -0.9762907
-        ordering: 207
-    Yn: 
-      metadata: 
-        scale: 7.154074e-18
-        bias: -0.01864709
-        ordering: 208
-    Ye: 
-      metadata: 
-        scale: 0.003166824
-        bias: -0.01864709
-        ordering: 209
-    Yx: 
-      metadata: 
-        scale: 0.02102178
-        bias: -0.3071955
-        ordering: 210
-    BTx: 
-      metadata: 
-        scale: 1.346439
-        bias: -3.118446
-        ordering: 211
-    MAXt: 
-      metadata: 
-        scale: 0.2419509
-        bias: -0.9853402
-        ordering: 212
-    BApressure: 
-      metadata: 
-        scale: 2.061877e-06
-        bias: -0.5213394
-        ordering: 213
-    tMINradius: 
-      metadata: 
-        scale: 1.392544
-        bias: -3.239921
-        ordering: 214
-    MINradius: 
-      metadata: 
-        scale: 0.06266253
-        bias: -1.384504
-        ordering: 215
-  images: 
-    metadata: 
-      dims: [64, 64]
-      channels: 4
-      scale: [29.258502, 858.26596, 100048.72, 4807207.0]
-      bias: [0.0, 0.0, 0.0, 0.0]
-      hwc: "chw"
-
-    (0.0, 0.0): 
-      0.0: 
-        emi: 
-          metadata: 
-            ordering: 300
-    (90.0, 0.0): 
-      0.0: 
-        emi: 
-          metadata: 
-            ordering: 301
-    (90.0, 78.0): 
-      0.0: 
-        emi: 
-          metadata: 
-            ordering: 302
-)ptext";
