--- src/data_readers/unit_test/test_data/hdf5_experiment_schema.yaml
@@ -1,20 +1,0 @@
-std::string const experiment_schema = R"ptext(
-inputs:
-  metadata:
-    pack: "datum"
-  shape_model_initial_modes:(4,3):
-
-outputs:
-  metadata:
-    pack: "datum"
-  scalars:
-    BWx:
-
-    BT:
-
-  images:
-    metadata:
-      coerce: "double"
-    (90.0, 0.0):
-      0.0:
-)ptext";
