--- src/data_readers/unit_test/test_data/hrrl_hdf5.prototext
@@ -1,93 +1,0 @@
-std::string const experiment_prototext = R"Alfred(
-  reader {
-    name: "hdf5_data_reader"
-    role: "train"
-    sample_list: "WORK_DIR/sample_list_fake.txt"
-    validation_percent: 0.1
-    percent_of_data_to_use: 1.0
-    data_schema_filename: "WORK_DIR/hrrl_data_schema.yaml"
-    experiment_schema_filename: "WORK_DIR/hrrl_experiment_schema.yaml"
-  }
-}
-model {
-  objective_function {
-    layer_term {
-      scale_factor: 1.0
-      layer: "layer6"
-    }
-  }
-  metric {
-    layer_metric {
-      layer: "layer6"
-      name: "mse"
-    }
-  }
-  num_epochs: 10
-  layer {
-    name: "layer1"
-    children: "layer2 layer3"
-    input {
-      target_mode: "regression"
-    }
-  }
-  layer {
-    name: "layer3"
-    parents: "layer1"
-    children: "layer6"
-    identity {
-    }
-  }
-  layer {
-    name: "layer2"
-    parents: "layer1"
-    children: "fc1"
-    identity {
-    }
-  }
-  layer {
-    name: "fc1"
-    parents: "layer2"
-    children: "pred"
-    data_layout: "data_parallel"
-    fully_connected {
-      num_neurons: 100
-      has_bias: true
-    }
-  }
-  layer {
-    name: "pred"
-    parents: "fc1"
-    children: "layer6"
-    data_layout: "data_parallel"
-    fully_connected {
-      num_neurons: 5
-      has_bias: true
-    }
-  }
-  layer {
-    name: "layer6"
-    parents: "layer3 pred"
-    mean_squared_error {
-    }
-  }
-  callback {
-    print {
-    }
-  }
-  callback {
-    timer {
-    }
-  }
-}
-optimizer {
-  adam {
-    learn_rate: 0.0001
-    beta1: 0.9
-    beta2: 0.99
-    eps: 1e-08
-  }
-}
-trainer {
-  mini_batch_size: 256
-}
-)Alfred";
